Pinning django-filter to 21.1 because it breaks PollingPlacesGeoJSONViewSet (aka /api/0.1/map/?election_id=33)
i.e. It's not picking up the election_id field in the request parameters any more, so it queries all polling places. This takes ages and causes the request to timeout.

Briefly tried a few things from the docs to make it work on 22.1, to no avail, so giving up for now and hoping it's a bug that's fixed.

https://www.django-rest-framework.org/api-guide/filtering/
https://github.com/carltongibson/django-filter