Persist the app's part of the Redux store into localStorage for a better UX
Mostly focussed around the user reloading the map and losing their search terms.

There's still small holes in this e.g. the user searching, going to the map (search term still shown), clearing the search term, and then going back in their browser and expecting to see the search terms on the map.

In that case, they'd have to go back far enough to get to the /<election_name>/search/location/... endpoint that did the search in the first place.