WIP code for when we were thinking about saving state in the URL *and* in the Redux store
Abandoned in favour of the Google Maps-style approach where the coords, zoom level, *and* search terms all go in the URL and that's that

We were thinking about storing map state in Redux so the user could return to the election in future and get back to exactly the same point, but realistically that's such a tiny number of users given how fleeting our usage is.

We were even starting to think about storing 'last updated' timestamps in Redux so we could discard stale information and how to handle the state per-election.

Not storing anything in Redux neatly side steps those questions and issues.