Well it seems to work OK on the About Page now
Could reproduce it locally in the iOS Simulator in XCode, but when I tested it was working using this dummy button on the AboutPage.

Tried removing 100dvh and .sentry-error-embed-wrapper locally, but still couldn't get it to work properly like it did for real.

If it occurs again, next steps might be debugging on my phone for real by connecting it to my laptop.